requirements.txt
pandas==2.1.4
numpy==1.26.2
streamlit==1.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6